refactor: move CommentNode
